+++ ingot-service/ingot-pms/ingot-pms-api/src/main/java/com/ingot/cloud/pms/api/model/domain/SysRoleAuthority.java
@@ -23,12 +23,11 @@ public class SysRoleAuthority extends BaseModel<SysRoleAuthority> {
     /**
      * 角色ID
      */
-    private Long roleId;
+    private Integer roleId;
 
     /**
      * 权限ID
      */
-    private Long authorityId;
-
+    private Integer authorityId;
 
 }
