+++ ingot-service/ingot-pms/ingot-pms-api/src/main/java/com/ingot/cloud/pms/api/model/domain/SysRoleDept.java
@@ -23,12 +23,11 @@ public class SysRoleDept extends BaseModel<SysRoleAuthority> {
     /**
      * 角色ID
      */
-    private Long roleId;
+    private Integer roleId;
 
     /**
      * 部门ID
      */
-    private Long deptId;
-
+    private Integer deptId;
 
 }
