+++ ingot-service/ingot-pms/ingot-pms-api/src/main/java/com/ingot/cloud/pms/api/model/domain/SysRoleMenu.java
@@ -23,12 +23,11 @@ public class SysRoleMenu extends BaseModel<SysRoleAuthority> {
     /**
      * 角色ID
      */
-    private Long roleId;
+    private Integer roleId;
 
     /**
      * 菜单ID
      */
-    private Long menuId;
-
+    private Integer menuId;
 
 }
