+++ ingot-service/ingot-pms/ingot-pms-api/src/main/java/com/ingot/cloud/pms/api/model/domain/SysRoleOauthClient.java
@@ -23,12 +23,11 @@ public class SysRoleOauthClient extends BaseModel<SysRoleAuthority> {
     /**
      * 角色ID
      */
-    private Long roleId;
+    private Integer roleId;
 
     /**
      * 客户端ID
      */
     private String clientId;
-
 
 }
