+++ ingot-service/ingot-pms/ingot-pms-api/src/main/java/com/ingot/cloud/pms/api/model/domain/SysRoleUser.java
@@ -23,12 +23,11 @@ public class SysRoleUser extends BaseModel<SysRoleAuthority> {
     /**
      * 角色ID
      */
-    private Long roleId;
+    private Integer roleId;
 
     /**
      * 用户ID
      */
-    private Long userId;
-
+    private Integer userId;
 
 }
